londiste.local_remove_table: resture DDL if still attached to table
Multi-TX COPY can happen on merge nodes.  If table
is removed during such copy, the DDL is removed from table
and put to 'dropped_ddl' column.

If dropped_ddl is not empty, the table is truncated and
DDL is restored.  Reason for truncate is that table is
unsynced anyway, so no point keeping the data.