londiste: don't filter EXECUTE or TRUNCATE events
They used to be filtered in merge nodes, but it does not
make sense anymore.

EXECUTE tracking is global now, not per-queue, thus filtering
is wrong now.

TRUNCATE does not seem to make sense in case of merged queue, but
filtering also does not make sense.  So let it pass down,
maybe there is scenariou where it is useful.