londiste resurrect: Get dead root back into cascade
In some cases the 'takeover --dead-root' will be used although
actual data on node is fine.  Eg. network downtime, lost power.
It would be good to have a way to bring it back to cascade
without need for full rebuild.

This command sync queue contents and re-subscribes node.
Any events found on old root that are not spread to rest of cascade
are dumped and deleted.  No actual table changes are made,
thus there must be external mechanism to avoid data conflicts.